# DataBase/apps/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: DataBase Gateway
  version: 0.1.0
  description: Thin API facade for the DataBase ecosystem.
servers:
  - url: http://127.0.0.1:18090
components:
  securitySchemes:
    DataBaseApiKey:
      type: apiKey
      in: header
      name: X-DataBase-Api-Key
      description: Optional. Required only when DATABASE_GATEWAY_AUTH_REQUIRED=true.
  schemas:
    RequestId:
      type: string
      description: Per-request correlation id returned in X-Request-Id.
    NullableDateTime:
      type:
        - string
        - "null"
      format: date-time
...
    VocabularyItem:
      type: object
      required:
        - id
        - content
        - type
        - category
        - note
        - created_at
        - updated_at
      properties:
        id:
          oneOf:
            - type: integer
            - type: string
        content:
          type: string
        type:
          type: string
        category:
          type: string
        note:
          type:
            - string
            - "null"
        created_at:
          $ref: "#/components/schemas/NullableDateTime"
        updated_at:
          $ref: "#/components/schemas/NullableDateTime"
    VocabularySearchResponse:
      type: object
      required:
        - query
        - count
        - items
        - requestId
      properties:
        query:
          type: string
        count:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/VocabularyItem"
        requestId:
          $ref: "#/components/schemas/RequestId"
...
    StylePack:
      type: object
      properties:
        version:
          type: string
          const: style-pack.v1
        query:
          type: string
        mode:
          type: string
          enum:
            - empty_query
            - style_reference_projection
        sourceIds:
          type: array
          items:
            type: string
        profiles:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              sourceId:
                anyOf:
                  - type: string
                  - type: "null"
              sourceTitle:
                type: string
              sourceAuthor:
                anyOf:
                  - type: string
                  - type: "null"
              sourceLocator:
                anyOf:
                  - type: string
                  - type: "null"
              summary:
                type: string
              sentenceLengthBand:
                type: string
              paragraphDensity:
                type: string
              progressionMoves:
                type: array
                items:
                  type: string
              rhetoricalMoves:
                type: array
                items:
                  type: string
              imageryClusters:
                type: array
                items:
                  type: string
              constraints:
                type: array
                items:
                  type: string
              tags:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    layer:
                      type: string
                    value:
                      type: string
                    description:
                      anyOf:
                        - type: string
                        - type: "null"
                  required:
                    - id
                    - layer
                    - value
                  additionalProperties: false
              searchScore:
                anyOf:
                  - type: number
                  - type: "null"
            required:
              - id
              - sourceTitle
              - summary
              - sentenceLengthBand
              - paragraphDensity
              - progressionMoves
              - rhetoricalMoves
              - imageryClusters
              - constraints
              - tags
            additionalProperties: false
        syntaxProfiles:
          type: array
          items:
            type: string
        rhetoricalMoves:
          type: array
          items:
            type: string
        imageryClusters:
          type: array
          items:
            type: string
        paragraphMoves:
          type: array
          items:
            type: string
        revisionPairs:
          default: []
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              sourceId:
                anyOf:
                  - type: string
                  - type: "null"
              sourceTitle:
                type: string
              sourceLocator:
                anyOf:
                  - type: string
                  - type: "null"
              issueType:
                type: string
              ruleId:
                type: string
              severity:
                type: string
              originalText:
                type: string
              revisedText:
                anyOf:
                  - type: string
                  - type: "null"
              reviewerEvidence:
                type: object
                properties:
                  badReason:
                    type: string
                  rewriteActions:
                    type: array
                    items:
                      type: string
                  forbiddenMoves:
                    type: array
                    items:
                      type: string
                  targetShape:
                    type: string
                required:
                  - badReason
                  - rewriteActions
                  - forbiddenMoves
                  - targetShape
                additionalProperties: false
              tags:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    layer:
                      type: string
                    value:
                      type: string
                    description:
                      anyOf:
                        - type: string
                        - type: "null"
                  required:
                    - id
                    - layer
                    - value
                  additionalProperties: false
              searchScore:
                anyOf:
                  - type: number
                  - type: "null"
            required:
              - id
              - sourceTitle
              - issueType
              - ruleId
              - severity
              - originalText
              - reviewerEvidence
              - tags
            additionalProperties: false
        constraints:
          type: array
          items:
            type: string
        counts:
          type: object
          properties:
            sources:
              type: number
            profiles:
              type: number
            syntaxProfiles:
              type: number
            rhetoricalMoves:
              type: number
            imageryClusters:
              type: number
            revisionPairs:
              default: 0
              type: number
          required:
            - sources
            - profiles
            - syntaxProfiles
            - rhetoricalMoves
            - imageryClusters
            - revisionPairs
          additionalProperties: false
        screening:
          type: object
          properties:
            version:
              type: string
              const: style-screening.v1
            requestedLimit:
              type: number
            sourceFilterIds:
              default: []
              type: array
              items:
                type: string
            selectedProfileCount:
              type: number
            sourceDiversityCount:
              type: number
            rankingSignals:
              type: array
              items:
                type: string
          required:
            - version
            - requestedLimit
            - sourceFilterIds
            - selectedProfileCount
            - sourceDiversityCount
            - rankingSignals
          additionalProperties: false
        requestId:
          type: string
      required:
        - version
        - query
        - mode
        - sourceIds
        - profiles
        - syntaxProfiles
        - rhetoricalMoves
        - imageryClusters
        - paragraphMoves
        - revisionPairs
        - constraints
        - counts
        - screening
        - requestId
      additionalProperties: false
...
paths:
...
  /vocabulary/search:
    get:
      operationId: searchVocabulary
      summary: Vocabulary search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
            maximum: 100
      responses:
        "200":
          description: Vocabulary matches.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VocabularySearchResponse"
        "401":
          description: Missing or invalid API key.
...
  /style/pack:
    get:
      operationId: getStylePack
      summary: Read DataBase style and syntax reference as StylePack
      description: Returns DataBase-owned derived style profiles for syntax, rhetoric, imagery, paragraph movement, and copyright boundaries. This is not factual evidence and does not expose reusable copyrighted sentences.
      parameters:
        - name: q
          in: query
          required: true
          description: Style query such as 句法 修辞 意象.
          schema:
            type: string
        - name: sourceIds
          in: query
          description: Comma/space separated DataBase source ids. Use this to constrain StylePack retrieval to a specific style source, for example book_kinkakuji_restricted_style.
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum style profiles to return.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 6
      responses:
        "200":
          description: DataBase-owned StylePack assembled from restricted style reference projections.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StylePack"
        "401":
          description: Missing or invalid API key.
...